requirements.txt
# Requirements for Algerian Dialect Fake News Detection Pipeline

# LangChain and LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20

# OpenAI API (used by langchain-openai)
openai>=1.12.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Pydantic for structured outputs
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Search functionality
# duckduckgo-search>=4.0.0
